Frotas - Django/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==4.2.3
djangorestframework==3.14.0
drf-spectacular==0.26.4
drf-yasg==1.21.7
exceptiongroup==1.2.0
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
pluggy==1.4.0
pytest==8.1.1
pytest-django==4.8.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.18.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.7.1
uritemplate==4.1.1
